tests/requirements.txt
pymssql>=2.2.0
pytest>=7.4.0
pytest-html>=4.0.0
pytest-json-report>=1.5.0
pyyaml>=6.0
jinja2>=3.1.0
colorama>=0.4.6
